Refactor within_tolerance to better represent use.
Function now takes a value and non-negative within value that the
abolute value of the input should be under. The tolerance for
isclose is not a float specific isclose parameter to handle only
the floating point equivalence issues. This is more flexible than
doing calculations in the function with the input values.